requirements.txt
python-dotenv
MLB-StatsAPI
tweepy
pytest
pytest-mock
Pillow
pytesseract
requests
